refactor: layout 구조 분리
